Mitigate undesireable side-effect of 8c5317f89151
Related discussion/issue:
- https://github.com/uBlockOrigin/uMatrix-issues/issues/156#issuecomment-494427094
- https://github.com/uBlockOrigin/uMatrix-issues/issues/155

Due to a Chromium issue about not providing the proper context
information, a negative side-effect was introduced with fix
to #155. This commit will force the originator of a network
request of type `main_frame` to be that of the request URL
itself, i.e. the originator of a top-level document request
is the requested document itself.